Updated lifetime to use UInt64

+++ Sources/RequestDL/Internals/Sources/Storage/Storage/Internals.Storage.swift
@@ -9,13 +9,13 @@ extension Internals {
     @RequestActor
     class Storage {
 
-        static let lifetime: Int = 5_000_000_000 * 60
+        static let lifetime: UInt64 = 5_000_000_000 * 60
         static let shared = Storage(lifetime: lifetime)
 
-        private let lifetime: Int
+        private let lifetime: UInt64
         private var table = [AnyHashable: Register]()
 
-        init(lifetime: Int) {
+        init(lifetime: UInt64) {
             self.lifetime = lifetime
             scheduleCleanup()
         }
